% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.fillReduce}
\alias{.fillReduce}
\title{Convert GRanges to Single Range}
\usage{
.fillReduce(gr, validate = TRUE)
}
\arguments{
\item{gr}{a \code{GRanges} with ranges to be merged.}

\item{validate}{logical determining whether entries should be checked for compatible
seqnames and strands.}
}
\value{
\code{GRanges} with single interval
}
\description{
Convert GRanges to Single Range
}
\details{
The validation assumes seqnames and strand are \code{Rle} objects.
}
