% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.mutateEach,CompressedGRangesList-method}
\alias{.mutateEach,CompressedGRangesList-method}
\title{Efficient Metadata Columns Mutation}
\usage{
\S4method{.mutateEach}{CompressedGRangesList}(grl, ...)
}
\arguments{
\item{grl}{a CompressedGRangesList}

\item{...}{named list of vectors to insert as metadata columns on each element
\code{GRanges}. Each vector length must match the length of the \code{GRangesList}.}
}
\value{
a \code{CompressedGRangesList} with all element \code{GRanges} updated
with supplied metadata columns
}
\description{
Efficient Metadata Columns Mutation
}
