% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{.propagateMap}
\alias{.propagateMap}
\title{Propagate Transcript Merge Map}
\usage{
.propagateMap(df, MAXITERS = 1000)
}
\arguments{
\item{df}{a \code{data.frame} with columns \code{tx_in} and \code{tx_out}}

\item{MAXITERS}{a numeric controlling the maximum number of iterations}
}
\value{
a converged \code{data.frame}, such that, \code{tx_out} is not
present in any \code{tx_in}
}
\description{
Propagate Transcript Merge Map
}
