% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.suppressTxDbGenomeWarning}
\alias{.suppressTxDbGenomeWarning}
\title{Suppress \code{txdbmaker} warning message}
\usage{
.suppressTxDbGenomeWarning(fn_call)
}
\arguments{
\item{fn_call}{an R expression that creates a TxDb object, expected a call to
\code{makeTxDbFromGRanges} or similar function.}
}
\value{
a \code{txdb} object representing a transcriptome, as returned by the
expression passed to \code{fn_call}.
}
\description{
Due to missing genome information in the input data,
\code{txdbmaker::makeTxDbFromGRanges} always returns a warning message
stating 'genome version information is not available for this TxDb object'.
This warning does not affect functionality and can be safely suppressed. This
function executes a TxDb-generating expression while suppressing only this
specific warning.
}
