% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncateTxome.R
\name{.clipTranscript}
\alias{.clipTranscript}
\title{Clip Transcript to Given Length}
\usage{
.clipTranscript(gr, maxTxLength)
}
\arguments{
\item{gr}{a \code{GRanges} object}

\item{maxTxLength}{a positive integer}
}
\value{
the clipped \code{GRanges} object
}
\description{
Internal function for operating on individual \code{GRanges}, where ranges
represent exons in a transcript. This is designed to be used in an
\code{*apply} function over a \code{GRangesList} object.
}
