% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{exportFASTA}
\alias{exportFASTA}
\title{Export Transcriptome as FASTA}
\usage{
exportFASTA(txdb, genome, file)
}
\arguments{
\item{txdb}{a \code{TxDb} object representing a transcriptome annotation}

\item{genome}{a \code{BSgenome} object from which to extract sequences}

\item{file}{output FASTA file}
}
\description{
Export Transcriptome as FASTA
}
