% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{exportFASTA}
\alias{exportFASTA}
\title{Export Transcriptome as FASTA}
\usage{
exportFASTA(txdb, genome, file)
}
\arguments{
\item{txdb}{a \code{TxDb} object representing a transcriptome annotation}

\item{genome}{a \code{BSgenome} object from which to extract sequences}

\item{file}{output FASTA file}
}
\value{
The \code{txdb} argument is invisibly returned.
}
\description{
Export Transcriptome as FASTA
}
\examples{
library(GenomicFeatures)
library(BSgenome)

## create TxDb using SGD genes for sacCer3 genome from UCSC
txdb <- makeTxDbFromUCSC("sacCer3", "sgdGene")

## retrieve sacCer3 sequence as BSgenome object
sacCer3 <- getBSgenome("sacCer3")

## restrict to 'chrI' transcripts (makes for briefer example runtime)
seqlevels(txdb) <- c("chrI")

## last 500 nts per tx
txdb_w500 <- truncateTxome(txdb)

exportFASTA(txdb_w500, sacCer3, "sacCer3.sgdGene.w500.fa")

}
