% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{exportGTF}
\alias{exportGTF}
\title{Export GTF}
\usage{
exportGTF(txdb, file, source = "txcutr")
}
\arguments{
\item{txdb}{transcriptome to be output}

\item{file}{location to write GTF}

\item{source}{a string to go in the \code{source} column}
}
\value{
The \code{txdb} argument is invisibly returned.
}
\description{
Exports a TxDb annotation to a GTF file
}
